Lets you take rounds out of a pocket ammo box (#2387)
## About The Pull Request

Exactly what it says on the tin. If an ammo box is in your pocket,
alt-clicking on it will, instead of taking the box, take a round out of
the box. ~~You can take it into your hand like you would a deck of cards
- by click-dragging.~~ With the action moved to alt-click, other
interactions of an ammo box should be unaffected.

## Why It's Good For The Game

Manual reload style points. I love reloading in the middle of a battle!

## Changelog

:cl:
tweak: You can now take individual rounds out of an ammo box in your
pocket by alt-clicking!
/:cl: