re:hair: baseline (#2464)
this is me asking nicely to not to

Resprites several hairs. I'm not even calling this part 1 because this
is barely only 4% of the hairs. The point of this is to establish a
consistent style to use for the rest. To make my job easier, i've cut
some hairs because there is 190 of them and a few removals shouldn't
hurt. I WILL remove more as I do the rest and find out redundant hairs/
hairs trying to do the same hairstyle but different quality.

I know remove not improve. I've done some new ones to make up for the
removals, but that's not for this PR.

As I know some removals will upset some people; below I will list the
hair cuts (heh), why, and a replacement

Curls. I actually respirited this under the impression it was a short
hime, until I found there was an actual short hime hairstyle. For
consistency I removed this one and put the resprite here. Use Short Hime
instead.

Hime 2. Barely different from Hime 1 and I don't know what it was trying
to change. Use Hime 1 instead.

Long Hair 3. It's literally long hair 2 but with a shitton of scribbles
on it. Use long hair 2 instead.

![image](https://github.com/shiptest-ss13/Shiptest/assets/58402542/443bf9cb-84e7-480f-a94c-ad968075d105)

![image](https://github.com/shiptest-ss13/Shiptest/assets/58402542/0c6218d3-ee56-4aa3-a658-5bd3c1f8b6c4)

![image](https://github.com/shiptest-ss13/Shiptest/assets/58402542/b7c6b58b-e41e-471f-9134-2057d6f80d95)

![image](https://github.com/shiptest-ss13/Shiptest/assets/58402542/ab02f40c-590c-4b51-95cf-2ceb1830a03f)

Not all the hairs redone, but the one's i'd like to showcase

:cl:
add: redoes about 9 hairs
del: about 3 hairs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->