Trickwine Hotfix (#2492)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

It took me 1-2 rounds the gather up a list of small issues and bugs
- [x] Stray pixel somewhere in the flask sprites.
- [x] Color of reagents doesn't match the sprites.
- [x] Typo in vintage flasks.
- [x] Prism wine needs buff: maybe give it a real reflect and make prism
wine actually reflect lasers.
- [x] The breakaway flask recipe should go in the drinks tab.
- [x] The distiller's icon state gets turned back to barrel when you
open it. Use a variable in the barrel to update icons so they acctually
get updated proper

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

fixing my mistakes and oversights is.. good c:
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
fix: fixed shitty trickwine code (who let this guy cook)
balance: buff to prismwine to make it useful
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->