Casually wearing your tactical turtleneck no longer makes it invisible (#2561)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

The tactical turtleneck didn't have a casual sprite variant for
digitigrades, this fixes that


![image](https://github.com/shiptest-ss13/Shiptest/assets/24857008/a6145e0b-7b4a-4ac8-a30b-d94c2b92b5ea)

![image](https://github.com/shiptest-ss13/Shiptest/assets/24857008/f8aad429-1e17-4590-9329-508797eafea5)


## Why It's Good For The Game

fixes #2325

## Changelog

🆑 
fix: The tactical turtleneck no longer tactically strips when worn
casually
/🆑

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->