Li Tieguai: Cybersun Edition (#2505)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

This PR reworks and updates the Li Tieguai's design into a Cybersun
Ship, as it was outlined in its lore, and what it was slated to become.

![2023-11-22 02 00
49](https://github.com/shiptest-ss13/Shiptest/assets/77556824/cc6a667a-e5dd-46ab-84fe-8351e050bf60)

![2023-11-22 02 00
50](https://github.com/shiptest-ss13/Shiptest/assets/77556824/68face8f-3bb9-4314-9cbc-3db5e92b9fa7)

(SHIP IS NOW FUCKING FLIPPED LET'S GO)

This PR also adds a new job, the Medical Director. It functions exactly
as how a CMO would, but for Cybersun instead!

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/5c93414c-4805-4b38-8ee7-e08ebde3c9ee)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

MEDICAL CYBERSUN SHIP IS FINALLY REAL OH MY GOD.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl: PositiveEntropy
add: The Li-Tieguai is now officially, a Syndicate ship!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->