Removes the (Independent) Lieutenant job (#2556)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
Removes the wildly inconsistent and unmaintained whitesands relic that
is the Lieutenant job (formerly known as the Solgov Representative. Yes,
that one). Pieces of its clothing were shoved into various ships or
outfits, and it was only ever used on the Shetland which is admittedly
very dated and in need of modernizing. This should not touch the LP
Lieutenant job or the Twinkleshine Lieutenant job.
The Shetland now has a First Officer instead of a Lieutenant, and now
has two separate lockers for the FO and Captain instead of just one for
the Lieutenant. That poor, poor captain, locker-less for so long..
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/d52b76b1-8821-41e6-94f3-b1d8cb997148)
This job has visible rust from its existence in the whitesands era and
tries and fails to fill in at what the First Officer job does better.
The independent captain/command resprite also broke its jumpsuit sprites
due to the type paths overlapping.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
del: Removes the Lieutenant job datum, locker and clothing
add: The Shetland now has a First Officer and lockers for both them and
the captain now
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->