Fixes Month Display (#2623)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Previously, the month display for IC/sector time was a month earlier
than it should have been. This also resulted in nothing displaying for
January because there is no month 0.

## Why It's Good For The Game
Time consistency and actually displaying the IC month

## Changelog

:cl:
fix: Sector time now actually shows the month of January.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->