Fix date calculation (again) (#2624)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
My math was all over the place. This makes it so that there's only one
day in year day and months don't start on the second. In my defense it's
kinda hard to test this.

## Why It's Good For The Game
Time consistency

## Changelog

:cl:
fix: IC/sector months now actually start on the first and there's only
one year day
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->