Fixes 2647 (#2648)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Arcane machine tried to smelt up to 10 sheets of material but the alloy
smelting proc didn't get a smelting number passed so it just spawned 1
sheet and called it a day.

## Why It's Good For The Game

fixes #2647

## Changelog

:cl:
fix: ore smelter no longer obliterates materials when smelting alloys
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Signed-off-by: Theos <theubernyan@gmail.com>