Fixes #2628 (#2674)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
Continuation of #2628, which was merged without a testmerge. This fixes
some ERT outfits, adds a Syndicate Inspector ERT (which had the outfit
defined but not the ERT), and fixes the independent sechud background.
Also messes with the CMM Inspector's icon.

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/a3ac82d6-e5c2-4028-9992-f2565604ac0b)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
blame zephyr

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: ACLF Syndicate Inspector ERT
fix: Independent background icons
fix: ERT Outfits not having access and proper names
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->