PGF Loadouts and Roles (#2700)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

There are now PGF loadouts for the game! The first few will only really
effect things once we've got a PGF ship in, but the rest can be used by
admins to make pre-ship PGF events!

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

This is just the next big step towards PGF integration. After this I
just need few more sprites (a small number compared to what's already
in), a ship, and to get off my butt and finalize that wiki page.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added PGF loadouts and roles. For playable on ships, there's
Crewman (Assistant), Engineer, Doctor, Marine, Marine Sergeant, and
Captain, and for adminspawn, there's all of the above fully kitted out
with equipment and the additional Marine Medic.
tweak: changed the name of one PGF item that I reconsidered the name of.
It didn't fit with the loadouts I was making.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: cml9988 <cml9988@g.rit.edu>