SRM Elder-class Montagnes' Vessel (Glaive Replacement) (#2554)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
Replaces the Glaive with the Elder-class.

![image](https://github.com/shiptest-ss13/Shiptest/assets/55075747/11ab1dd8-0227-4d6f-aa28-6fe41e9742bc)

![image](https://github.com/shiptest-ss13/Shiptest/assets/55075747/3fc417f1-6780-4a0e-b6b0-c7b83bea218a)
![elder class
real](https://github.com/shiptest-ss13/Shiptest/assets/55075747/eca32c3a-a37d-4aad-838b-05dca54daef4)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
The Glaive is old and a maptainer has requested I maintain my maps and
replace or revamp it. I chose the former.
This version gives more attention to the non-glade sections and makes
the glade less janky. This version also puts more of a focus on the
non-mining and ruin running aspects of the old ship. The ashie med,
botany, animals, and ashwines.
You cannot actually fish in the ship. No I will not be making it so you
can. Not only would that not make sense, but it would require porting
over normal fishability from the planetary turfs onto the shipturf
version of water. Which I'm just not doing.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:phoaly
add: Added the Elder-class
del: Removed the Glaive-class
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->