Twink Tweaks: The Second (#2536)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/e0558629-b3e4-46d7-9695-dafe63a62a06)
Updates the armory, cryogenics, bar, bridge, and the equipment of the
Twinkleshine. Eliminates the vendor and voucher system in favor of a
more controlled stockpile by the Lieutenant. Crew now spawn in with
barely any equipment, excluding the voicechanger mask, which now works
(for real trust me), with some other stuff there and there. The Captain
is now a Flotilla Admiral, to reflect the importance of a Twinkleshine.
Adds the Redshield Officer.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
I should be making ruins.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: Syndicate Cryopod.
tweak: The Twinkleshine has gotten a facelift.
tweak: The Syndicate has started assigning Flotilla Admirals and
Redshield Officers to Twinkleshines.
tweak: Crew on the Twinkleshine now have to get their stuff from their
lockers.
tweak: Modified the starting text to reflect the lack of telecomms on
most ships.
fix: Access on the Twinkleshine.
fix: Syndicate jobs should show up properly on SecHUDs.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: meem <75212565+meemofcourse@users.noreply.github.com>