Inteq-ening (#2748)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
Self-explanatory title! This was originally a PR that was going to
merely change how the Vanguard and the Master At Arms looked like, but
it mutated into a near full-blown visual rework!

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/a4fc0ac1-9bce-400c-a763-9f089f8e41fb)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Inteq is no longer rotting.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl: PositiveEntropy
imageadd: Inteq has now received a visual overhaul!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->