Small interview QoL improvements (#2743)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Ports tgstation/tgstation#59621 which allows us to embed links in
interview questions. Also fixes an annoying config verification error
that didn't mean anything.

## Why It's Good For The Game
Allows us to link rules and lore in interviews, so people are asked to
read them.

## Changelog

:cl: bobbahbrown, Mark Suckerberg
admin: Links can now be embedded in interview questions to get players
to read them.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Bobbahbrown <bobbahbrown@gmail.com>