Kills combat knife crate and replaces it with survival knives (#2736)
## About The Pull Request
As the name says!!!!!
## Why It's Good For The Game
Holdover from TG, massively overpriced for today's shiptest and combat
knives are being kept as a higher-tier loot item.
## Changelog

:cl:

balance: Changes combat knives in cargo to survival knives and cheapens
their crate.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->