Adds survival knives to the Colossus, Talos and Vaquero. (#2735) As the name says!
It's kind of fucked only the Valor gets them.

:cl:
add: Survival knives to every inteq ship.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->