Adds zoom to the GAL, SSG-04 and EBR. (#2810)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds the ability to zoom on the scope for a few rifles.

## Why It's Good For The Game
Encourages acting as a squad, buffs them a little. little silly that the
scope is only decorative

## Changelog

:cl:
add: The SSG-04, GAL and EBR can now use their scopes!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->