Updates paper signature text size (#2827)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
it updates the pixel size of signatures used in paper

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/95ac90ae-2b62-4cae-9845-70c750b7c629)

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/ede0e93e-7b21-410b-ab7e-fc92233cf57b)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
they are barely readable otherwise

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: Modified the pixel size of signatures in paper to 15px
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->