Spooky's Objects Resprites (#2780)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
A number of assorted resprites for all sorts of objects.

I didn't test this, so I'll let tests run and ensure it compiles fine.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

![image](https://github.com/shiptest-ss13/Shiptest/assets/34109002/d3aaf065-2d58-4923-83c2-fdcd9767581c)

![image](https://github.com/shiptest-ss13/Shiptest/assets/34109002/649effea-0ec0-4b4b-823d-5aa8d2b768d3)

![image](https://github.com/shiptest-ss13/Shiptest/assets/34109002/72670766-a3f2-443e-94bf-f7463482e45f)

![image](https://github.com/shiptest-ss13/Shiptest/assets/34109002/f98b64a2-6bcd-4ac6-beb4-a1a3b539446c)
(Don't worry about the duplicate iconstates I fixed those in the actual
DMI)

![image](https://github.com/shiptest-ss13/Shiptest/assets/34109002/41fa6e79-b158-42b2-843f-de0b1799f0ab)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl: Spooky
tweak: Assorted resprites.
imageadd: A bunch of re-done sprites for various objects. Notably:
Sofas, filing cabinets, reagent tanks.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->