Modernises the Meta-class and changes its faction to NT (#2566)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
The Meta was a comfy little whiteship in dire need of fixing. It had
tinyfans, light fixture frames and an uncomfortably cramped kitchen and
botany area. This PR remaps it while keeping its soul intact, making it
an old Nanotrasen vessel instead of an Indie.

NEW META (NO AREAS):
![StrongDMM-2023-12-06 00 54
04](https://github.com/shiptest-ss13/Shiptest/assets/95449138/bd542fdb-8de4-462e-bc48-c7f8e49e2b41)

NEW META (AREAS):
![StrongDMM-2023-12-06 00 54
00](https://github.com/shiptest-ss13/Shiptest/assets/95449138/a6ea0332-18f4-466b-a65b-19dd1d08da1f)

NEW META (INGAME):

![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/d2eff008-052a-4bd9-99ab-7d14170ec156)

OLD META (NO AREAS):
![StrongDMM-2023-12-05 17 42
10](https://github.com/shiptest-ss13/Shiptest/assets/95449138/d1a8c051-3cb4-4846-a357-731526c8a7e9)
I don't want this pretty good ship to be removed for being old and
having tinyfans. Having another NT ship that fills a logistics/cargo
niche is also pretty great.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: Remapped the Meta-class and changed its faction to NT
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->